+++ ricaun.NUnit/Models/TestModel.cs
@@ -54,6 +54,8 @@
             var result = Skipped ? "Skipped" : Success ? "Passed" : "Failed";
             var name = string.IsNullOrEmpty(Alias) ? Name : Alias != Name ? $"{Alias}[{Name}]" : Alias;
             var message = Message;
+            if (string.IsNullOrEmpty(message))
+                message = Console;
             //var message = (!Message.Contains("Exception:")) ? Message : Message.Substring(0, Message.IndexOf("Exception:") + "Exception".Length);
             return $"{name}\t {result}\t {message}".Replace("\n", " ").Replace("\r", " ");
         }