+++ ricaun.NUnit/Services/TestAssemblyService.cs
@@ -70,6 +70,23 @@ namespace ricaun.NUnit.Services
         }
 
         /// <summary>
+        /// Get TypeInstance for to run tests
+        /// </summary>
+        /// <param name="useExported"></param>
+        /// <returns></returns>
+        private IEnumerable<TypeInstance> GetTestTypeInstances(bool useExported = false)
+        {
+            var types = GetTestTypes(useExported);
+            foreach (var type in types)
+            {
+                foreach (var testFixture in GetTestFixtureAttributes(type))
+                {
+                    yield return new TypeInstance(type, testFixture.Arguments);
+                }
+            }
+        }
+
+        /// <summary>
         /// Get the types from the specified assembly.
         /// </summary>
         /// <param name="useExported">A flag indicating whether to use exported types only.</param>
@@ -84,13 +101,13 @@ namespace ricaun.NUnit.Services
         /// </summary>
         /// <param name="useExported">A flag indicating whether to use exported types only.</param>
         /// <returns></returns>
-        public IReadOnlyDictionary<Type, MethodInfo[]> GetTestDictionaryTypeMethods(bool useExported = false)
+        public IReadOnlyDictionary<TypeInstance, MethodInfo[]> GetTestDictionaryTypeMethods(bool useExported = false)
         {
-            var result = new Dictionary<Type, MethodInfo[]>();
-            var types = GetTestTypes(useExported);
+            var result = new Dictionary<TypeInstance, MethodInfo[]>();
+            var types = GetTestTypeInstances(useExported);
             foreach (var type in types)
             {
-                var methods = GetMethodWithTestAttribute(type).ToArray();
+                var methods = GetMethodWithTestAttribute(type.Type).ToArray();
                 if (methods.Any())
                 {
                     result.Add(type, methods);
@@ -108,7 +125,7 @@ namespace ricaun.NUnit.Services
         {
             foreach (var typeMethod in GetTestDictionaryTypeMethods(useExported))
             {
-                Type type = typeMethod.Key;
+                var type = typeMethod.Key;
                 MethodInfo[] methods = typeMethod.Value;
                 foreach (MethodInfo method in methods)
                 {
@@ -128,7 +145,7 @@ namespace ricaun.NUnit.Services
         public IEnumerable<TestTypeModel> RunTests(bool useExported = false)
         {
             var result = new List<TestTypeModel>();
-            var types = GetTestTypes(useExported).Where(AnyMethodWithTestAttributeAndFilter);
+            var types = GetTestTypeInstances(useExported).Where(AnyMethodWithTestAttributeAndFilter);
 
             foreach (var type in types)
             {
@@ -137,7 +154,7 @@ namespace ricaun.NUnit.Services
                     var testTypeModel = new TestTypeModel();
                     try
                     {
-                        using (var test = new TestService(type, parameters))
+                        using (var test = new TestService(type, this.parameters))
                         {
                             testTypeModel = test.TestInstance();
                         }